requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.1

# Document Processing
PyMuPDF==1.23.8
python-docx==1.1.0

# Text Analysis
textblob==0.17.1
spacy==3.7.2
textstat==0.7.3
sentence-transformers==2.3.1
datasketch==1.6.4
nltk==3.8.1

# File Storage
minio==7.2.3

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Export
reportlab==4.0.9

# Frontend (simple)
jinja2==3.1.3
